requirements.txt
fastapi
uvicorn
requests
wand
apscheduler
liquidctl
pydantic-settings
psutil